move images to separate folder
